test reorganization: analysis and coordinates
- new subdirectory analysis/
- split the unwieldy test_analysis.py into separate
  files and removed superfluous imports (and got rid of
  the awful 'from numpy.tests import *': use only classes/
  functions that are actually needed)
- also moved test_mol2 and test_pdbqt to the coordinates/
  directory
- part of work on reorganizing tests #466
- todo: fixing tempfile handling #440
        density tests #410
        issues with PSA tests (#487 and #440)
