+++ testsuite/MDAnalysisTests/coordinates/test_mol2.py
